Moved threshold functions to subpackage
New threshold subpackage contains:
* plantcv.threshold.binary
* plantcv.threshold.gaussian
* plantcv.threshold.mean
* plantcv.threshold.otsu
* plantcv.threshold.triangle

This change reduced code redundancy and simplified function calls in some cases.

Some input variables were normalized also:
* img => gray_img
* maxValue => max_value
* object_type now defaults to "light"
