+++ docs/binary_threshold.md
@@ -3,15 +3,16 @@
 Creates a binary image from a gray image based on the threshold values. 
 The object target can be specified as dark or light.
 
-**plantcv.binary_threshold(*img, threshold, maxValue, object_type*)**
+**plantcv.threshold.binary(*gray_img, threshold, max_value, object_type="light"*)**
 
-**returns** thresholded image
+**returns** thresholded/binary image
 
 - **Parameters:**
-    - img - grayscale img object
-    - threshold - threshold value (0-255)
-    - maxValue - value to apply above threshold (255 = white)
-    - objecttype - 'light' or 'dark', is target image light or dark?
+    - gray_img - Grayscale image data
+    - threshold - Threshold value (0-255)
+    - max_value - Value to apply above threshold (255 = white)
+    - object_type - "light" or "dark" (default: "light"). If object is lighter than the background then standard 
+    thresholding is done. If object is darker than the background then inverse thresholding is done
 - **Context:**
     - Used to help differentiate plant and background
 - **Example use:**
@@ -35,7 +36,7 @@ from plantcv import plantcv as pcv
 pcv.params.debug = "print"
 
 # Create binary image from a gray image based on threshold values. Targeting light objects in the image.
-threshold_light = pcv.binary_threshold(img, 36, 255, 'light')
+threshold_light = pcv.threshold.binary(gray_img, 36, 255, 'light')
 ```
 
 **Thresholded image**
@@ -50,7 +51,7 @@ from plantcv import plantcv as pcv
 pcv.params.debug = "print"
 
 # Create binary image from a gray image based on threshold values. Targeting dark objects in the image.
-threshold_dark = pcv.binary_threshold(img, 36, 255, 'dark')
+threshold_dark = pcv.threshold.binary(gray_img, 36, 255, 'dark')
 ```
 
 **Thresholded image (inverse)**
