+++ docs/triangle_threshold.md
@@ -2,14 +2,15 @@
 
 Creates a binary image from a gray image using adaptive thresholding.
 
-**plantcv.triangle_auto_threshold(*img, maxvalue, object_type, xstep=1*)**
+**plantcv.threshold.triangle(*gray_img, max_value, object_type="light", xstep=1*)**
 
-**returns** thresholded image
+**returns** thresholded/binary image
 
 - **Parameters:**
-    - img - grayscale img object
-    - maxValue - value to apply above threshold (255 = white)
-    - objecttype - 'light' or 'dark', is target image light or dark?
+    - gray_img - Grayscale image data
+    - max_value - Value to apply above threshold (255 = white)
+    - object_type - "light" or "dark" (default: "light"). If object is lighter than the background then standard 
+    thresholding is done. If object is darker than the background then inverse thresholding is done
     - xstep - value to move along x-axis to determine the points from which to calculate distance
               recommended to start at 1 and change if needed)
 - **Context:**
@@ -28,7 +29,7 @@ from plantcv import plantcv as pcv
 pcv.params.debug = "print"
 
 # Create binary image from a gray image based
-thresholded = pcv.triangle_auto_threshold(img, 255,'light', xstep=10)
+thresholded = pcv.threshold.triangle(gray_img, 255, 'light', xstep=10)
 ```
 
 **Triangle Auto-Thresholded image (xstep=10)**
@@ -43,7 +44,7 @@ from plantcv import plantcv as pcv
 pcv.params.debug = "print"
 
 # Create binary image from a gray image based 
-thresholded = pcv.triangle_auto_threshold(img, 255,'light', xstep=1)
+thresholded = pcv.threshold.triangle(gray_img, 255, 'light', xstep=1)
 ```
 
 **Triangle Auto-Thresholded image (xstep=1)**
