+++ plantcv/plantcv/threshold/__init__.py
@@ -1,0 +1,7 @@
+from plantcv.plantcv.threshold.threshold_methods import binary
+from plantcv.plantcv.threshold.threshold_methods import gaussian
+from plantcv.plantcv.threshold.threshold_methods import mean
+from plantcv.plantcv.threshold.threshold_methods import otsu
+from plantcv.plantcv.threshold.threshold_methods import triangle
+
+__all__ = ["binary", "gaussian", "mean", "otsu", "triangle"]
